ingestion-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
confluent-kafka==2.3.0
protobuf==4.25.1
